Data Enumeration and Interpreter
Added the data enumeration to look up PID name and number of bytes that is supposed to be returned.  Added test string to Main to run through interpretter.  If string doesn't end with a carriage return, the next one is appended to it, in case of split response (might be removed, and just reject the response).